[PB-250] refactor global search component
