+++ src/app/core/components/Navbar/EmptyState.tsx
@@ -1,0 +1,22 @@
+import {} from '@phosphor-icons/react';
+import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
+import iconService from 'app/drive/services/icon.service';
+
+const EmptyState = (): JSX.Element => {
+  const { translate } = useTranslationContext();
+  const PdfIcon = iconService.getItemIcon(false, 'pdf');
+  const FolderIcon = iconService.getItemIcon(true);
+
+  return (
+    <div className="flex h-full flex-col items-center justify-center">
+      <div className="relative h-20 w-28">
+        <FolderIcon className="absolute top-0 left-11 h-16 w-16 rotate-10 transform drop-shadow-soft filter" />
+        <PdfIcon className="absolute top-0 left-2 h-16 w-16 rotate-10- transform drop-shadow-soft filter" />
+      </div>
+      <p className="text-xl font-medium text-gray-100">{translate('general.searchBar.emptyState.title')}</p>
+      <p className="text-sm font-normal text-gray-60">{translate('general.searchBar.emptyState.subtitle')}</p>
+    </div>
+  );
+};
+
+export default EmptyState;
